==> workflows/01_building_dataset_from_Sherwood2020.Rmd <==
---
title: "Building dataset of ECS values from Sherwood 2020"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Goal 

Build and save data set of ECS values from different ECS configuration distributions as provided by S20 supplemental information.

# Building data set

Here we are building ECS data sets for each of the evidence configurations of interest. We are using data from [Sherwood et al. 2020](https://agupubs.onlinelibrary.wiley.com/doi/full/10.1029/2019RG000678) henceforth referred to as S20. The data was drawn from the supplemental information in S20 and represents ECS percentile estimates from likelihood distributions quantified in S20 for each of the five evidence configurations (each containing different combinations of lines of evidence). There are other configurations available in S20 beyond the five we have chosen here. The percentile estimates reported in this data represent the 5th, 10th, 17th, 20th, 25th, 50th, 75th, 80th, 83rd, 80th, and 95th percentile, as well as the mode and the mean. 

```{r}
ecs_data_list <- list(
  "Baseline" = c(
    2.2550000000000026,
    2.4050000000000082,
    2.5549999999999997,
    2.615000000000002,
    2.7050000000000054,
    3.105000000000011,
    3.625000000000007,
    3.7749999999999986,
    3.884999999999998,
    4.2350000000000065,
    4.685000000000009
  ),
  "No_Process" = c(
    2.045000000000009,
    2.2150000000000105,
    2.395000000000003,
    2.4550000000000054,
    2.565000000000005,
    3.0549999999999997,
    3.7250000000000014,
    3.9250000000000043,
    4.065000000000005,
    4.535000000000004,
    5.155000000000008
  ),
  "No_Historical" = c(
    2.0050000000000026,
    2.164999999999999,
    2.32500000000001,
    2.384999999999998,
    2.4750000000000014,
    2.9050000000000082,
    3.4550000000000054,
    3.625000000000007,
    3.7350000000000065,
    4.115000000000002,
    4.605000000000011
  ),
  "No_Paleoclimate" = c(
    2.2450000000000117,
    2.4250000000000043,
    2.605000000000011,
    2.6750000000000043,
    2.7850000000000037,
    3.32500000000001,
    4.134999999999998,
    4.395000000000003,
    4.595000000000006,
    5.285000000000004,
    6.355000000000011 
  ),
  "Baseline_Emergent_constraints" = c(
    2.2550000000000026,
    2.4050000000000082,
    2.5549999999999997,
    2.615000000000002,
    2.7050000000000054,
    3.105000000000011,
    3.625000000000007,
    3.7749999999999986,
    3.884999999999998,
    4.2350000000000065,
    4.685000000000009
  )
)
```

Write data frame and store in the `data` directory.

```{r}
saveRDS(ecs_data_list, "data/ecs_prior_data_S20.RDS")
```
