==> workflows/03_plotting_ecs_samples_S1.Rmd <==
---
title: "Plotting ECS distirbutions from estimated gamma distributions"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Goal 

Produce a figure with the ECS samples. This may serve as a good supplemental figure for the paper to show the distribution shapes for the evidence configurations from S20, based on gamma distribution assumption. 

# Produce figure

We produce PDF curves in a panel using the color palette and then save the figure in the `figures` directory.

This code will will produce the publication level figure. A simple visualization is available in `02_fitting_andsampling_ecs_distributions.Rmd`.

```{r}
ecs_sample_list <- readRDS("data/ecs_samples_plot_data.RDS")

# convert to dataframe
ecs_df_S1 <- data.frame(
  
  value = unlist(ecs_sample_list),
  scenario = rep(names(ecs_sample_list), each = n),
  row.names = NULL
  
)

# recode evidence configurations for aesthetics
ecs_df_S1_fig_data <- recode_scenarios(ecs_df_S1)

# convert the evidence configuration to factor with the facet order we want
ecs_df_S1_fig_data$scenario <- factor(ecs_df_S1_fig_data$scenario, levels = scenario_order)

# plot with ggplot"
ggplot() +
  geom_density(data = ecs_df_S1_fig_data,
               aes(x = value, 
                   color = scenario,
                   fill = scenario),
               linewidth = 0.75, 
               alpha = 0.2, 
               bw = 0.3) +
  scale_color_manual(values = ECS_COLORS, 
                     guide = "none") +
  scale_fill_manual(values = ECS_COLORS, 
                    guide = "none") +
  scale_x_continuous(breaks = seq(from = 0, to = 9, 
                                  by = 1)) +
  labs(x = "Equilibrium Climate Sensitivity Value",
       y = "Density") +
  facet_wrap(~ scenario) +
  theme_light() +
  theme(legend.position = "bottom",
        strip.text = element_text(size = 10, color = "black"), 
        strip.background = element_rect(fill = "white"))

ggsave("figures/S1_fig_ECS_distributions.png",
       device = "png",
       width = 7,
       height = 5,
       units = "in",
       dpi = "print")
```
