==> workflows/04_running_the_model_computing_probabilities.Rmd <==
---
title: "Using ECS sample to run matilda"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Goal

The goal of this script is to run Matilda with each of the ECS distributions we sampled prior. 

# Using ECS samples to run Matilda

We use ECS values sampled from the estimated parametric distributions from S20 to propagate the varying levels of uncertainty associated with evidence configurations to probabilistic climate projections. This provides an opportunity to better understand how different ECS evidence configurations affect temperature trajectories from a simple carbon cycle climate model. 

We use the SSP2-4.5 emission scenario to run the model with the perturbed ECS samples.

```{r}
# read in ecs samples as a list
ecs_sample_list <- readRDS("data/initial_ecs_samples_from_gamma_dist.RDS")

# read in scenario input file (ini)
ini_245 <- system.file("input/hector_ssp245.ini", package = "hector")

```

The scenario input file is used to initiate a `core` environment for Hector. Hector is the simple carbon-cycle climate model that is the engine behind Matilda's probabilistic projection framework. More details about Hector and its usage, visit the [Hector GitHub page](https://jgcri.github.io/hector/).

```{r}
# initiate model core
core <- newcore(ini_245)

```

The result will be a new core object that can will be a required input to run the model. 

# Generate values for other model parameters

Matilda works by running Hector multiple times to build a perturbed parameter ensemble, thus applying parameter uncertainty to model outputs. We need to produce parameter values to accompany the ECS values we sampled in previous steps of the workflow.

Parameter sets are generated in Matilda using `generate_params`. We use this function to produce `n` initial parameter sets (`init_params`). In this analysis I do not think I am going to run samples of the other parameters. Instead we can isolate the behavior of Hector to different ECS distributions by using Hector defaults for all parameters aside from ECS.

```{r, eval=FALSE}
# set seed for reproducible result
set.seed(123)

# sample size (should match ECS sample)
n = 10000

# generate parameters
init_params <- generate_params(core = core, draws = n)

```

The result will be a new data frame object with 15,000 samples for 6 parameters.

*NOTE*: This data frame includes a column for `ECS`. These are samples drawn from the default prior distribution in Matilda, not the distributions selected for this analysis. 

We replace the default generated `ECS` values with the values we sampled from S20 distributions. This gives us a set of model parameters that are identical except for the `ECS` column, which isolates the impact of propagating `ECS` uncertainty through the model. 

```{r, eval=FALSE}
# create a list of parameter data frames based on ECS samples in ecs_sample_list
parameter_list <- lapply(ecs_sample_list, function(ECS) {

  # copy init_params
  params_no_ecs <- init_params

  # remove the ECS column from the parameter data frame
  params_no_ecs$ECS <- NULL

  # add sampled S20 ecs values
  cbind(ECS, params_no_ecs)

})

```

The result is a list of parameter sets named after the evidence configuration used to produce the ECS values. 

# Run the model 

We use each of the parameter sets in `parameter_list` to run the model. This produces a single Hector run for each of the 15,000 parameter sets per each ECS evidence configuration (15,000 x 5 = 75,000 total model runs).

Parallel computing on the local machine is used to make this process as efficient as possible. 

```{r}
# split the ecs samples into chunks for each 'worker'
parameter_chunks <- lapply(parameter_list, function(df) {
  
  split(df, 1:1000)
  
})

# detect cores 
detectCores()

# initiate a cluster
cl <- makeCluster(detectCores() - 1)

# export required functions and objects to run the model
clusterExport(cl, c("parameter_chunks",
                    "ini_245",
                    "newcore",
                    "reset",
                    "iterate_model"))
# start time
start_time <- Sys.time()

# run the model with parLapply
model_result <- parLapply(cl, parameter_chunks, function(evidence_config) {
  
  # initialize a model core for each loop iteration
  core <- newcore(ini_245)
  
  # run the model for each parameter chunk
  result_list <- lapply(evidence_config, function(chunk) {
    
    reset(core)
    
    iterate_model(core = core,
                  params = chunk,
                  save_years = 1800:2100,
                  save_vars = c("gmst", 
                                "global_tas",
                                "CO2_concentration", 
                                "ocean_uptake"))
  })

  # ensure correct run_number added to each parameter chunk
  for (i in 2:length(result_list)) {

    # calculate the max value of the previous element in result_list
    max_run_number <- max(result_list[[i - 1]]$run_number)

    # add the max value of the previous element to the run_number of the current
    # element to get a run_number that is continuous from the previous element.
    result_list[[i]]$run_number <- result_list[[i]]$run_number + max_run_number
  }
  
  # bind parameter_chunks
  result <- do.call(rbind, result_list)
  
  return(result)
})

# Stop time
run_time <- Sys.time() - start_time
print(run_time)

# stop the cluster
stopCluster(cl)

# save the result
saveRDS(model_result, "data/raw_unweighted_model_results.RDS")
```
This can take ~9 hours to run 60,000 model iterations (10,000 runs x 6 scenarios).

The result is `model_result` a list of Matilda outputs, one for each `ECS` distribution based on a specific evidence configuration. Each result in the list contains 10,000 Hector runs using `ECS` samples from prior steps of the analysis. We saved the years 1800:2100 for the variables `global_tas`, `gmst`, `co2_concentration`, and `ocean_uptake`. 

Add data needed for time-varying error -- for `gmst` criterion.

```{r}
# HadCRU5 temperature data
gmst_criteria_data <- read.csv("data-raw/gmst_anomaly_hadcrut5.csv")
# remove confidence bounds in df - leaving only year and mean anomaly cols
gmst_criteria_data$Lower.confidence.limit..2.5.. <- NULL
gmst_criteria_data$Upper.confidence.limit..97.5.. <- NULL
gmst_criteria_data <- gmst_criteria_data %>% rename("value" = "Anomaly..deg.C.",
                                                    "year" = "Time")

# Temperature anomaly error vector
gmst_error <- read.csv("data-raw/annual_gmst_SD_quant.csv")
temp_error_vec <- gmst_error$value
```

Plot temperature data with time varying uncertainty error as a sanity check:

```{r}
# for plotting add error values to criteria data 
gmst_criteria_data$error <- gmst_error$value
ggplot() +
  geom_line(data = gmst_criteria_data, 
            aes(x = year, 
                y = value)) +
  geom_ribbon(data = gmst_criteria_data, 
              aes(x = year,           
                  ymin = value - error, 
                  ymax = value + error), 
              alpha = 0.2)             

```

Create new temperature criterion:
```{r}
# create a GMST criterion
gmst_criterion <- new_criterion("gmst", years = gmst_criteria_data$year, obs_values = gmst_criteria_data$value)

```

*The Matilda provided gmst criterion only uses a time frame 1961-2023. We create the new criterion to use the timeframe of 1850-2024.* 

Create criterion using ocean carbon uptake from the Global Carbon Project:

```{r}
# read in ocean uptake data 
ocean_sink <- read.csv("data-raw/annual_ocean_c_uptake.csv", stringsAsFactors = F)

# create a new ocean c uptake criterion
ocean_uptake_criterion <- new_criterion("ocean_uptake", years = ocean_sink$year, obs_values = ocean_sink$value)
```

With these criterion created, we can weight ensemble members. Here, we will weight using temperature, CO2 concentration, and ocean carbon uptake criterion. Weights are the likelihood of an ensemble member based on the its averaged RMSE with regard to each to the 3 criterion and are represented as normalized probabilities. In this analysis the influence of the criterion on the final ensemble weight is equal. We use the `score_bayesian` as our scoring function with a default `sigma = sd(criterion_data)`.

```{r}
# weight models using observed co2 and temp - store in a list
model_weights <- lapply(model_result, function(df) {
  
  # produce weights based on co2
  weights_co2 = score_runs(df,
                           criterion = criterion_co2_obs(),
                           score_function = score_bayesian,
                           sigma = 0.12)
  weights_co2 = na.omit(weights_co2)

  # produce weights based on temp
  weights_temp = score_runs(df,
                            criterion = gmst_criterion,
                            score_function = score_bayesian, 
                            sigma = temp_error_vec)
  weights_temp = na.omit(weights_temp)

  # produce weights based on ocean carbon uptake
   weights_ocean_uptake = score_runs(df,
                                    criterion = ocean_uptake_criterion,
                                    score_function = score_bayesian)
   weights_ocean_uptake = na.omit(weights_ocean_uptake)

   # store in a list
   weights_list = list(weights_co2, weights_temp, weights_ocean_uptake)

   # compute multi-criteria weights
   mc_weights = multi_criteria_weighting(weights_list, c(0.15, 0.8, 0.05))

  return(mc_weights)
  
})

## Check model_weights for names
names(model_weights)

## model_weights before culling constraint
saveRDS(model_weights, "data/pre_culled_model_weights.RDS")
```

Filter out ensemble members with near-zero weight:

```{r}
# filter out ensemble members that do not meet the minimum weight constraint (> 1e-6)
constrained_weights <- lapply(model_weights, function(df) {
  
  filtered_result <- subset(df, mc_weight > 1e-6)
  
  return(filtered_result)
  
  })

```

Constrained weights can be merged with the constrained results. This would produce a list (based on ECS scenario) of the constrained model ensemble and the assigned weights for each run. However, because some of the models have been filtered out during the constraint, need to re-normalize so weights sum to 1. This way we can still use the resulting ensembles to compute metrics and probabilities accurately.

```{r}
weighted_ensemble <- Map(function(a, b) {
  
  merged <- merge(a, b, by = "run_number")
  
  return(merged)

  }, constrained_weights, model_result)

```

Sanity check to ensure names are correct and re-weighted ensemble sums to 1.0:
```{r}
#sanity check names 
names(weighted_ensemble)

## Re-weight to sum to 1
norm_weights_ensemble <- lapply(names(weighted_ensemble), function(df_name) {

  # make copy of data
  df <- weighted_ensemble[[df_name]]

  # calculate the total weight for unique run_numbers
  total_weight <- sum(df$mc_weight[!duplicated(df$run_number)])

  # normalize the weight values
  df$norm_weight <- df$mc_weight / total_weight

  # add "scenario" name
  df$scenario <- df_name

  return(df)

})
```

Sanity check that element names are preserved and final weights sum to 1.0:
```{r}
# make sure element names are preserved
names(norm_weights_ensemble) <- names(weighted_ensemble)

# Verify the final weights sum to 1.0
sapply(norm_weights_ensemble, function(df) {
  
  sum <- sum(df$norm_weight[!duplicated(df$run_number)])

  print(sum)  
})
```

```{r}
# saver normalized weighted ensemble
saveRDS(norm_weights_ensemble, "data/weighted_ensemble.RDS")

```

The result of the weighting step is a list of ECS configurations, each with a data frame containing weights for each ensemble member that was not culled from the temperature or minimum weight constraint (1e-6). Initial weights computed with each criterion individually were used to run `multi_criteria_weighting` which computes a weight taking into account all weighting criterion used.

# Computing Metrics

We will compute temperature metrics for median 2080-2100 warming relative to 1995-2014 reference. 

Normalize temperature to the 1995-2014 reference period. Do this by extracting `global_tas` as warming data then normalize.

```{r}
global_tas_data <- lapply(norm_weights_ensemble, function(df){
  
  subset(df,
         variable == "global_tas" 
         & year > 1849 
         & year < 2101)
  })

gsat_results <- lapply(global_tas_data, function(df){
    
  # Filter data for the reference period
  normalized_values <- normalize_dat(df, 
                                     ref_start = 1995, 
                                     ref_end = 2014)
  
  # adding this column to each df
  df$value <- normalized_values
  
  return(df)
  
})

saveRDS(gsat_results, "data/weighted_gsat_ensemble.RDS")
```

Define the metric we are interested in calculating.

```{r}
# define metric of interest - end of century (eoc) warming
eoc_metric <- new_metric(var = "global_tas", years = 2081:2100, op = median)

```

We use the newly defined metric object to compute median 2081-2100 warming for each ECS configuration scenario. Additionally, we add scenario names and merge weights for each model. 

```{r}
# Build data frame of metric results
eoc_warming_results <- lapply(gsat_results, function(df){

  # compute metrics for each df in the weighted model results list using eoc_warming metric object
  metric_df <- metric_calc(df, eoc_metric)
  
  return(metric_df)
})

# Merge with model weights to get a data frame with weighted metrics
eoc_gsat_metrics <- Map(function(a, b){
  
  b <- b[!duplicated(b$run_number), ]
  
  merged <- merge(a, b[, c("run_number", "norm_weight")], by = "run_number")
  
  return(merged)
  
}, eoc_warming_results, norm_weights_ensemble)

```

Sanity check to verify the normaized metric weights sum to 1.0:
```{r}
# Apply function to calculate and print sum of the specified column
# Verify the final normalized weights
sapply(eoc_gsat_metrics, function(df) {
  
  sum <- sum(df$norm_weight)

  print(sum)  
})

```

Save weighted metric data:
```{r}
# save the result for future visualization
saveRDS(eoc_gsat_metrics, "data/weighted_gsat_metrics.RDS")

```

# Computing Probabilities

We compute probabilities using the likelihood weights and the warming metrics produced for each ECS scenario. The probability calculation sums weights (which total to 1.0) as warming metrics are grouped into bins we define. Here, bins represent ranges of warming that could potentially occur at the end of the century. In this way, the sum of the weights for each bin (warming range) is proportional to the probability of that warming range occurring according to the models response to parameter uncertainty. This step is where model weights become particularly important because a higher weight (representing closer alignment to historical data) will have a larger influence on the total probability of a warming range than an unskilled model (low likelihood based on alignment with historical data).

To compute probabilities we call `prob_calc` for each of the data frame in the `weighted_eoc_warming_results` list.

```{r}
# define bins that represent the warming ranges of interest
temp_ranges <- c(0, 1, 1.5, 2, 2.5, 3, 3.5, 4, 4.5, Inf)

# computing probabilities 
temp_probability <- lapply(names(eoc_gsat_metrics), function(df_name){
  
  # copy data based on element name
  df <- eoc_gsat_metrics[[df_name]]
  
  # run prob_calc
  prob_result <- prob_calc(df$metric_result,
                           bins = temp_ranges,
                           scores = df$norm_weight)
  
  # add scenario column 
  prob_result$scenario <- df_name
  
  return(prob_result)
})

# copy element names again
names(temp_probability) <- names(eoc_gsat_metrics)

# Save full data frame for future visualization
saveRDS(temp_probability, "data/temp_probability_results.RDS")

```

The result from this code is a list of data frames (one for each ECS scenario) that contain the weighted temperature range probabilities.

The last thing that we will produce with this set of code is a probability result where the probability is computed for a sequence of temperature ranges. To complete this we will use temperature bins defined as a sequence.
