==> workflows/05_running_chi_square_on_probabilities.Rmd <==
---
title: "Chi-square test on warming probabilities of ECS scenarios"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Goal

The goal of this script is to run a chi-squared test to determine if there is an association between ECS scenario and the likelihood of specific temperature ranges. For example, this can inform whether one ECS configuration leads to a higher likelihood of warming by a given temperature range compared to another ECS configuration. In this analysis, high chi-squared values (significant p-value) suggest that certain ECS configurations influence the likelihood of the temperature range outcome. Conversely, low chi-square statistic (non-significant p-value) suggests there is not enough evidence to conclude that ECS configuration affects the likelihood of a particular warming range. Thus, this analysis can help discern whether ECS configurations have a notable impact on projected temperature outcomes and how the ECS parameter affects future temperature projections.

```{r}
library(reshape2)
library(tidyverse)
```

# Data

This analysis requires the `probability_results`. We load this data from the `data` directory.

```{r}
# load probability data
options(scipen = 999)

prob_data <- readRDS("data/temp_probability_results.RDS")

prob_data <- do.call(rbind, prob_data) %>% 
  mutate(scenario = factor(scenario),
         probability = probability * 100) %>% 
  rename(temperature_range = "bins")

```

Upon loading the data we remove `rownames` and convert `scenario` to a factor.

# Run Chi-square test

H0: There is no significant association between ECS configurations and the likelihood of different end-of-century warming ranges. The probability of temperature ranges is independent of the ECS range.

H1: There is a significant association between ECS configurations and the likelihood of different end-of-century warming ranges. Certain ECS configurations influence the likelihood of certain end-of-century warming occurring. 

```{r}
# Reshape the data into a contingency table format
contingency_table <- dcast(prob_data, 
                           temperature_range ~ scenario, 
                           value.var = "probability")
contingency_table[is.na(contingency_table)] <- 0

# Compute chi-squared results
result2 <- chisq.test(contingency_table[-1])

# split the data by temperature range 
table_split <- split(contingency_table, 
                     contingency_table$temperature_range)

# Run chi-square test on each temperature range
result <- lapply(table_split, function(table) {
  
  # Isolate the probability values in the contingency table
  values <- table[-1]
  
  # Compute chi-squared results
  result <- chisq.test(values)
  
  # Build a results data frame
  stats <- data.frame(X.squared = result$statistic,
                      df = result$parameter,
                      p.value_chi = result$p.value)
  
  return(stats)
})

stats_table <- do.call(rbind, result)

```

The warning produced is concerning. Probably due to the way I am using chi-squared test (splitting the contingency table). The only time the analysis does not produce the warning is when I don't divide into temperature ranges and simulate p-value. The result is largely uninformative I would say. Tells us that likelihood of certain warming ranges are not impacted by ECS configuration, only on that is close to significant is warming range of 3.5-4 C, where `No_Paleoclimate` may be associated with a higher likelihood of occurrence. 

```{r}
# Visualize the data with a bar plot
library(ggplot2)

# Assuming your data is named 'df'
# Reshape the data from wide to long format
df_long <- pivot_longer(contingency_table, -temperature_range, names_to = "scenario", values_to = "probability")

# Create the bar plot
ggplot(prob_data, aes(x = temperature_range, y = probability, color = scenario, group = scenario)) +
  geom_point(size = 3)+
  geom_line(linewidth = 0.7) +
  labs(title = "Probability Distribution by Scenario",
       x = "Warming Range (C)",
       y = "Probability") +
  theme_light() +
  theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
  scale_color_manual(values = c("Baseline" = "#003466", 
                                "No_Process" = "#00a9cf",
                                "No_Historical" = "#550307",
                                "No_Paleoclimate" = "#EBCC2A",
                                "Baseline_Emergent_constraints" = "#F21A00"))
```


# Testing out KS test

```{r}
# Load the data
warming_metrics <- readRDS("data/weighted_gsat_metrics.RDS")

# resampled warming metrics with weights
resampled_metrics <- lapply(warming_metrics, function(df) {
  
  #new samples 
  new_sample <- sample(df$metric_result, size = 10000, replace = T, prob = df$norm_weght)
  
  return(new_sample)
  
})

# Extract names of distributions
names_distributions <- names(warming_metrics)

# Generate all unique pairs of distribution indices and names
pairs <- combn(seq_along(resampled_metrics), 2, simplify = FALSE)

# Function to compute KS test for a pair of distributions
compute_ks_test <- function(pair) {
  # Extract distribution names and data
  name1 <- names_distributions[pair[1]]
  name2 <- names_distributions[pair[2]]
  dist1 <- metric_results[[pair[1]]]
  dist2 <- metric_results[[pair[2]]]
  
  # Perform the KS test
  ks_results <- ks.test(dist1, dist2)
  
  # Return a list with pair names and test results
  return(list(pair = c(name1, name2), ks_test = ks_results))
}

# Perform pairwise KS tests
ks_test_results <- lapply(pairs, compute_ks_test)

# Print results
print(ks_test_results)

```

```{r}
Baseline <- data.frame(metric = resampled_metrics$Baseline, scenario = "Baseline")
IPCC <- data.frame(metric = resampled_metrics$IPCC, scenario = "IPCC")
No_Process <- data.frame(metric = resampled_metrics$No_Process, scenario = "No_Process")
No_Historical <- data.frame(metric = resampled_metrics$No_Historical, scenario = "No_Historical")
No_Paleoclimate <- data.frame(metric = resampled_metrics$No_Paleoclimate, scenario = "No_Paleoclimate")
baseline_EC <- data.frame(metric = resampled_metrics$Baseline_Emergent_constraints, scenario = "Baseline_Emergent_constraints")

# Combine data
combined_data <- rbind(Baseline, IPCC, No_Process, No_Historical, No_Paleoclimate, baseline_EC)

# recoding scenario names for the plot
cdf_df <- recode_scenarios(combined_data)

# Order factor levels for plot 
scenario_order <- 
  c("IPCC AR6",
    "Baseline",
    "No Process",
    "No Historical",
    "No Paleoclimate",
    "Baseline + Emergent constraints")

# Plot CDFs
CDF_temp <- ggplot(cdf_df, 
                   aes(x = metric, color = scenario)) +
  stat_ecdf(geom = "step", size = 1) +
  scale_color_manual(values = ECS_COLORS) +
  labs(title = "CDFs of Two Distributions", x = "Value", y = "CDF") +
  theme_light()

ggsave("figures/cdf_figure.png",
       CDF_temp,
       device = "png", 
       width = 12, 
       height = 8, 
       units = "in", 
       dpi = "print")
```
