==> workflows/05_running_ks_test.Rmd <==
---
title: "Chi-square test on warming probabilities of ECS scenarios"
output: github_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Goal

The goal of this script is to run a KS test to determine if there is a significant difference among all temperature distributions for data computed using each ECS configuration.

# Run KS test 

```{r}
# Load the data
warming_metrics <- readRDS("data/weighted_gsat_metrics.RDS")

# resampled warming metrics with weights
metric_values <- lapply(warming_metrics, function(df) {
  
  # extract values 
  sample <- df$metric_result
  
  return(sample)
  
})

# Extract names of distributions
names_distributions <- names(warming_metrics)

# Generate all unique pairs of distribution indices and names
pairs <- combn(seq_along(metric_values), 2, simplify = FALSE)

# Perform pairwise KS tests
ks_test_results <- lapply(pairs, compute_ks_test)

# Print results
print(ks_test_results)

```
# Plot CDF curves to visualize differences in distributions 

```{r}
# Build data frames of metric values and Scenario names
Baseline <- data.frame(metric = metric_values$Baseline, scenario = "Baseline")
IPCC <- data.frame(metric = metric_values$IPCC, scenario = "IPCC AR6*")
No_Process <- data.frame(metric = metric_values$No_Process, scenario = "No_Process")
No_Historical <- data.frame(metric = metric_values$No_Historical, scenario = "No_Historical")
No_Paleoclimate <- data.frame(metric = metric_values$No_Paleoclimate, scenario = "No_Paleoclimate")
baseline_EC <- data.frame(metric = metric_values$Baseline_Emergent_constraints, scenario = "Baseline_Emergent_constraints")

# Combine data
combined_data <- rbind(Baseline, IPCC, No_Process, No_Historical, No_Paleoclimate, baseline_EC)

# recoding scenario names for the plot
cdf_df <- recode_scenarios(combined_data)

cdf_df$scenario <- factor(cdf_df$scenario, levels = names(ECS_COLORS))

# Plot CDFs
CDF_temp <- ggplot(cdf_df, 
                   aes(x = metric, 
                       color = scenario, 
                       linetype = scenario)) +
  geom_line(stat = "ecdf", 
            linewidth = 1,) +
  scale_x_continuous(limits = c(0, 4.5), expand = c(0, 0)) +
  scale_color_manual(values = ECS_COLORS, name = "ECS configurations") +
  scale_linetype_manual(values = LINE_TYPE, guide = "none") +
  labs(x = expression(paste("Future Warming (GSAT) relative to 1995-2014 (", degree, "C)")), 
       y = "Cumulative Density") +
  theme_light(base_size = 16) +
  theme(axis.title = element_text(size = 20), 
        axis.text = element_text(size = 18),
        legend.title = element_text(size = 20),
        legend.text = element_text(size = 18))
CDF_temp

# Adding line type guide
CDF_fig <- CDF_temp + guides(color = guide_legend(override.aes = list(color = ECS_COLORS, 
                                                                      linetype = LINE_TYPE)))
CDF_fig

ggsave("figures/fig_4_CDF.png",
       CDF_fig,
       device = "png", 
       width = 12, 
       height = 8, 
       units = "in", 
       dpi = "print")
```
