==> workflows/06_plotting_results.Rmd <==
---
title: "Producing Figures"
author: "Joe Brown"
date: "2024-06-19"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

# Goal

The goal of this script is to produce figures to visualize results of ECS evidence analysis.

# Weighted ensemble

```{r}
# load weighted ensemble data
gsat_results <- readRDS("data/weighted_gsat_ensemble.RDS")

# combine results into a data frame for plotting
gsat_df <- do.call(rbind, gsat_results)

# Re-code scenario names, factorize and order factor levels by `scenario_order`
gsat_plot_data <- recode_scenarios(gsat_df)
gsat_plot_data$scenario <- factor(gsat_plot_data$scenario, levels = scenario_order)

ggplot(data = gsat_plot_data) +
  geom_line(aes(x = year, 
                y = value, 
                group = run_number, 
                color = norm_weight, 
                alpha = norm_weight), 
            linewidth= 0.05) +
  scale_color_gradient(low = "lightblue", high = "dodgerblue4", name = "Weights") +
  scale_alpha_continuous(range(c(0,1))) +
  labs(x = "Years", y = expression(paste("Global Temperature Anomaly relative to 1995-2014 (", degree, "C)"))) +
  theme_light() +
  guides(alpha = "none") +
  facet_wrap(~ scenario) +
  theme(strip.text = element_text(size = 12, color = "black"), 
        strip.background = element_rect(fill = "white"), 
        axis.text = element_text(size = 12), 
        axis.title.x = element_text(size = 16), 
        axis.title.y = element_text(size = 14))
  
ggsave("figures/S2_fig_gsat_ensemble.png",
       device = "png",
       width = 28,
       height = 15,
       units = "cm",
       dpi = "print")
```

# Median temperature and confidence interval

Compute the weighted median and confidence intervals.

```{r}
# median (CI) warming calculation
warming_projection <- gsat_df %>% 
  group_by(year, scenario) %>% 
  summarize(
    median = weighted.quantile(x = value, w = norm_weight, probs = 0.5), 
    ci_5 = weighted.quantile(x = value, w = norm_weight, probs = 0.05),
    ci_16 = weighted.quantile(x = value, w = norm_weight, probs = 0.16), 
    ci_84 = weighted.quantile(x = value, w = norm_weight, probs = 0.84), 
    ci_95 = weighted.quantile(x = value, w = norm_weight, probs = 0.95),
    ci_66 = weighted.quantile(x = value, w = norm_weight, probs = 0.66), 
    ci_33 = weighted.quantile(x = value, w = norm_weight, probs = 0.33),
    .groups = "drop")

```

Editing data for plotting temperature projections:

```{r}
# recoding scenario names for the plot
warming_projection_data <- recode_scenarios(warming_projection)

# convert the evidence configuration to factor with the facet order we want
warming_projection$scenario <- factor(warming_projection$scenario, levels = scenario_order)

```

Construct plot: 

```{r}
temp_projection_plot <- 
  ggplot(data = subset(warming_projection_data, 
                       year > 2023)) +
  geom_line(aes(x = year, 
                y = median, 
                color = scenario), 
            linewidth = 0.75) +
  scale_color_manual(values = ECS_COLORS) +
  geom_ribbon(aes(x = year, 
                  ymin = ci_5, 
                  ymax = ci_95, 
                  fill = scenario, 
                  color = scenario), 
              alpha = 0.3, 
              linetype = "blank") +
  scale_fill_manual(values = ECS_COLORS) +
  labs(x = "Year", 
       y = expression(paste("Future Warming (GSAT) relative to 1995-2014 (", degree, "C)"))) +
    facet_wrap(~scenario) +
  theme_light() +
  theme(legend.position = "none", 
        axis.text.x = element_text(size = 22, angle = 35)) +
  theme(axis.text = element_text(size = 22), 
        axis.text.x = element_text(size = 24, vjust = 0.9),
        axis.title.x = element_blank(), 
        axis.title.y = element_text(size = 24), 
        strip.text = element_text(size = 22, color = "black"), 
        strip.background = element_rect(fill = "white"))

temp_projection_plot
```
Subset data for historical period:
```{r}
historic_tas_subset <- subset(gsat_df, 
                                year > 1849 &
                                year < 2024)

# recoding scenario names for the plot
historic_tas <- recode_scenarios(historic_tas_subset)

historic_plot_data <- historic_tas %>% 
  group_by(scenario, year) %>% 
  summarize(
    median = weighted.quantile(value, w = norm_weight, probs = 0.5),
    lower_ci = weighted.quantile(value, w = norm_weight, probs = 0.05),
    upper_ci = weighted.quantile(value, w = norm_weight, probs = 0.95),
    .groups = "drop")

```

Normalize historical temperature to reference period:
```{r}
# convert the evidence configuration to factor with the facet order we want
historic_plot_data$scenario <- factor(historic_plot_data$scenario, levels = scenario_order)

# normalizing historical data to the reference period
temp_hist <- read.csv("data-raw/annual_gmst_normalized.csv")
reference_historical <- normalize_historical(temp_hist, 1995:2014)
```

Plot temperature projection median line with ribbon for uncertainty range:
```{r}
warming_projection_plot <- 
  temp_projection_plot +
    geom_point(data = reference_historical, 
             aes(x = year, 
                 y = value), 
             color = SSP_COLORS[[8]], 
             size = 0.3) + 
  geom_line(data = historic_plot_data, 
            aes(x = year, 
                y = median), 
            color = SSP_COLORS[[9]],
            linewidth = 0.7) +
  geom_ribbon(data = historic_plot_data, 
              aes(x = year, 
                  ymin = lower_ci, 
                  ymax = upper_ci), 
              fill = SSP_COLORS[[9]],
              alpha = 0.2) +
  theme_light() +
  theme(axis.text = element_text(size = 22), 
        axis.text.x = element_text(size = 21, vjust = 0.9))


ggsave("figures/fig_3a_temp_warming_projections.png",
       warming_projection_plot,
       device = "png", 
       units = "in",
       width = 15,
       height = 10,
       dpi = "print")
```

# Temperature probability

Plotting temperature probabilities for each of the ECS evidence configurations.

Loading data:
```{r}
temp_probability <- readRDS("data/temp_probability_results.RDS")

temp_probability_df <- do.call(rbind, temp_probability)
row.names(temp_probability_df) <- NULL

temp_probability_df <- temp_probability_df %>% 
  complete(scenario, bins) %>% 
  mutate(bins = recode(bins, !!!bins_mapping))

temp_probability_df[is.na(temp_probability_df)] <- 0
```

```{r}
# recoding scenario names for the plot
temp_prob_plot_data <- recode_scenarios(temp_probability_df)

# order factor levels
temp_prob_plot_data$scenario <- factor(temp_prob_plot_data$scenario, 
                                       levels = prob_scenario_order)
```

The `bang bang bang` operator (`!!!`) expands vectors/lists into individual arguments. Here, this operator allows us to replace the `bins` factor levels according to the vectors in `bins_mapping`. 

Constructing plot:
```{r}
probability_plot <- 
  ggplot(data = temp_prob_plot_data, 
         aes(fill = bins, 
             x = scenario, 
             y = probability)) +
  geom_bar(position = position_fill(reverse = T), 
           stat = "identity", 
           width = 0.6) +
  scale_y_continuous(breaks = seq(0.0, 1.0, 0.1)) +
  scale_fill_manual(
    values = TEMP_PROBABILITY_COLORS,
    labels = bins_mapping,
    name = "Warming") +
  labs(y = "Probability", 
       x = NULL) +
  coord_flip() +
  theme_light(base_size = 20) +
  theme(legend.position = "right", 
        axis.text = element_text(size = 22), 
        axis.text.x = element_text(size = 22),
        axis.title.x = element_text(size = 24),
        legend.text = element_text(size = 22))

probability_plot

```

Save the plot:
```{r}
# save the probability plot
ggsave("figures/fig_3c_temperature_probability_plot.png", 
       probability_plot, 
       device = "png", 
       width = 14,
       height = 7, 
       units = "in", 
       dpi = "print")

```

# GSAT stats plot

This plot shows the spread, distribution shape and some descriptive statistics of the end-of-century `gsat` metrics.

```{r}
gsat_metrics <- readRDS("data/weighted_gsat_metrics.RDS")

# Add the 'scenario' column to each dataframe in the list
gsat_metrics <- lapply(names(gsat_metrics), function(scenario_name) {
  df <- gsat_metrics[[scenario_name]]
  df$scenario <- scenario_name
  return(df)
})

gsat_metrics_df <- do.call(rbind, gsat_metrics)
gsat_metrics_df$scenario <- factor(gsat_metrics_df$scenario)

gsat_metrics_stats <- data_summary(gsat_metrics_df)

```

```{r}
gsat_metrics_df <- recode_scenarios(gsat_metrics_df)
gsat_metrics_stats <- recode_scenarios(gsat_metrics_stats)

```

Constructing plot:
```{r}
gsat_stats_plot <- 
  ggplot() +
  geom_violin(data = gsat_metrics_df, 
              aes(x = scenario, 
                  y = metric_result, 
                  color = scenario, 
                  fill = scenario),
              linewidth = 0.7,
              alpha = 0.1,
              trim = FALSE) +
  geom_errorbar(data = gsat_metrics_stats, 
                aes(x = scenario, 
                    y = median, 
                    ymin = lower, 
                    ymax = upper, 
                    color = scenario), 
                width = 0.2, 
                linewidth = 0.8) +
  geom_point(data = gsat_metrics_stats,
             aes(x = scenario, 
                 y = median, 
                 color = scenario), 
             size = 6) +
  scale_color_manual(values = ECS_COLORS, name = "ECS configuration", guide = "none") +
  scale_fill_manual(values = ECS_COLORS, guide = "none") +
  ylim(0, 6.2) +
  geom_label(data = gsat_metrics_stats, 
             aes(x = scenario, 
                 y = upper + 0.5, 
                 label = paste(round(median, 2), "\n[", round(lower, 2), ",", round(upper, 2), "]")), 
             size = 6) +
  theme_light(base_size = 20) +
  theme(axis.title.y = element_blank(),
        axis.title.x = element_blank(),
        axis.text.x = element_text(angle = 35, vjust = 1.05, hjust = 1),
        axis.text = element_text(size = 20))

gsat_stats_plot
```

```{r}
ggsave("figures/fig_3b_temperature_metric_distribution.png",
       gsat_stats_plot,
       device = "png", 
       width = 7, 
       height = 6, 
       units = "in", 
       dpi = "print")
```

Adding `warming_projection_plot` and `gsat_stats_plot` to the same paneled figure:
```{r}
# label 
Panel_a <- warming_projection_plot + ggtitle("A") + 
  theme(plot.title = element_text(size = 24, face = "bold"), 
        axis.text.x = element_text(size = 22, angle = 35, vjust = 0.9, hjust = 1), 
        legend.position = "none", 
        axis.text = element_text(size = 22), 
        axis.title.x = element_blank(), 
        axis.title.y = element_text(size = 24), 
        strip.text = element_text(size = 22, color = "black"), 
        strip.background = element_rect(fill = "white")) 

Panel_b <- gsat_stats_plot + ggtitle("B") + theme(plot.title = element_text(size = 24, face = "bold"))

Panel_c <- probability_plot + ggtitle("C") + theme(plot.title = element_text(size = 24, face = "bold"))

top_row <- ggarrange(Panel_a, Panel_b, ncol = 2, nrow = 1, 
                                widths = c(3, 2), heights = c(4, 2))

combined_temp_plot <- ggarrange(top_row, Panel_c, ncol = 1, nrow = 2)

combined_temp_plot

ggsave("figures/fig_3_temperature_combined.png",
       combined_temp_plot,
       device = "png", 
       width = 25, 
       height = 18, 
       units = "in", 
       dpi = "print")
```

# Plotting ECS posterior distributions

Combine ECS samples with model weights:
```{r}
# Bring in ECS values
ecs_values <- readRDS("data/ecs_samples_plot_data.RDS")

# Add run_number column
ecs_priors <- lapply(ecs_values, function(df){
  
  df$run_number <- 1:10000

  return(df)
  
  })

# Bring in model_weights
model_weights <- readRDS("data/pre_culled_model_weights.RDS")

# Merge model weights with ECS priors using run_number
weighted_ECS <- Map(function(a, b){
  
  data_merge <- merge(a, b, by = "run_number")
  
  return(data_merge)
  
  }, ecs_priors, model_weights)

```

Constrain weighted_ECS and "re-weight" to match the weighted_ensemble:
```{r}
constrained_ecs_weights <- lapply(weighted_ECS, function(df) {
  
  filtered_result <- subset(df, mc_weight > 1e-6)
  
  return(filtered_result)
  
  })

# normalize weights
norm_weighted_ECS <- lapply(names(constrained_ecs_weights), function(df_name) {

  # make copy of data
  df <- constrained_ecs_weights[[df_name]]

  # calculate the total weight for unique run_numbers
  total_weight <- sum(df$mc_weight[!duplicated(df$run_number)])

  # normalize the weight values
  df$norm_weight <- df$mc_weight / total_weight

  # add "scenario" name
  df$scenario <- df_name

  return(df)

})

# check to ensure weights sum to 1
sapply(norm_weighted_ECS, function(df) {
  
  sum <- sum(df$norm_weight[!duplicated(df$run_number)])

  print(sum)  
  
})

names(norm_weighted_ECS) <- names(constrained_ecs_weights)
```

Computing kde_values:
```{r}
# use lapply to loop across data frames in scenario_list to compute KDE values for each
kde_values <- lapply(names(norm_weighted_ECS), function(df_name){
  
  df <- norm_weighted_ECS[[df_name]]
  
  # use that df to compute density estimates weight by `weights`
  density_values <- density(df$ECS, weights = df$norm_weight, bw = 0.5)
  
  # build data frame with coordinates we want to plot
  density_estimate <- data.frame(scenario = df_name,
                                  value = density_values$x,
                                  density = density_values$y)
  
  return(density_estimate)
})

# binding df for plotting
kde_df <- do.call(rbind, kde_values)
```

Now we can construct the plots. First we need to define breaks we want to represent to define the shading of intervals in the plot:

```{r}
# recoding scenario names for the plot
kde_data <- recode_scenarios(kde_df)

# order scenario levels
  
# outlining desired breaks
breaks <- c(0, 2.0, 2.5, 4.0, 5.0, Inf)

# legend lables
legend_labels <- c("0 to 2.0 C",
                   "2.0 to 2.5 C",
                   "2.5 to 4.0 C",
                   "4.0 to 5.0 C",
                   "5.0+ C")

# create intervals and assign colors
kde_df$interval <- cut(kde_df$value, 
                       breaks = breaks, 
                       include.lowest = TRUE, 
                       right = TRUE)

kde_df$alpha <- case_when(
  kde_df$interval == "[0,2]" ~ 0.3,  # Low likelihood (more transparent)
  kde_df$interval == "(2,2.5]" ~ 0.8, # Very likely
  kde_df$interval == "(2.5,4]" ~ 1.0, # Likely
  kde_df$interval == "(4,5]" ~ 0.8,   # Very likely
  kde_df$interval == "(5,Inf]" ~ 0.3, # Low likelihood (more transparent)
)


```

Plot:
```{r}
ECS_posterior <- 
  ggplot(kde_df, 
         aes(x = value, 
             y = density,
             fill = interval,
             alpha = alpha)) +
  geom_area(color = "darkgrey") +
  scale_fill_manual(values = ECS_POSTERIOR_COLORS, 
                    labels = legend_labels, 
                    name = "ECS values") +
  scale_alpha_identity() +
  scale_x_continuous(breaks = seq(0, 9, by = 1), # set the scale and labels of the x-axis
                     limits = c(0, 9), # set x-axis limits
                     expand = c(0, 0.1)) + # limit white space
  labs(x = expression(paste("Equilibrium Climate Sensitivity (", degree, "C)")), 
       y = "Density") +
  facet_wrap(~scenario) +
  theme_light(base_size = 18) +
  theme(legend.position = "none",
        axis.title.y = element_text(size = 22), 
        axis.title.x = element_text(size = 22), 
        strip.text = element_text(size = 20, color = "black"), 
        strip.background = element_rect(fill = "white"))
ECS_posterior

```

```{r}
ggsave("figures/fig_2_ECS_posterior_distribution.png",
       ECS_posterior,
       device = "png", 
       width = 16, 
       height = 8, 
       units = "in", 
       dpi = "print")
```

# EXTRA PLOT -- Posterior ECS distirbutions

```{r}
norm_weighted_ecs_df <- do.call(rbind, norm_weighted_ECS)

ecs_data_summary <- function(data){

  ecs_stats <-
    data %>%
    group_by(scenario) %>%
    summarize(
      median = weighted.quantile(ECS, w = norm_weight, probs = 0.5),
      vlikely_lower = weighted.quantile(ECS, w = norm_weight, probs = 0.05),
      likely_lower = weighted.quantile(ECS, w = norm_weight, probs = 0.33), 
      likely_upper = weighted.quantile(ECS, w = norm_weight, probs = 0.66),
      vlikely_upper = weighted.quantile(ECS, w = norm_weight, probs = 0.95),
      .groups = "drop")

  return(ecs_stats)

}

ecs_stats <- ecs_data_summary(norm_weighted_ecs_df)
```

```{r}
norm_weighted_ecs_df <- recode_scenarios(norm_weighted_ecs_df)
norm_weighted_ecs_df <- recode_scenarios(norm_weighted_ecs_df)

ecs_stats <- recode_scenarios(ecs_stats)
```

Constructing plot:
```{r}
ecs_stats_plot <- 
  ggplot() +
    geom_jitter(data = norm_weighted_ecs_df, 
              aes(x = scenario, 
                  y = ECS,
                  color = scenario),
              stat = "identity",
              width = 0.20,
              alpha = 0.05, 
              size = 0.7) +
  geom_violin(data = norm_weighted_ecs_df, 
              aes(x = scenario, 
                  y = ECS, 
                  color = scenario),
              linewidth = 0.7,
              alpha = 0.0,
              trim = FALSE) +
  geom_errorbar(data = ecs_stats, 
                aes(x = scenario, 
                    y = median, 
                    ymin = likely_lower, 
                    ymax = likely_upper, 
                    color = scenario), 
                width = 0.2, 
                linewidth = 0.8) +
  geom_point(data = ecs_stats,
             aes(x = scenario, 
                 y = median, 
                 color = scenario), 
             size = 6) +
  scale_color_manual(values = ECS_COLORS, name = "ECS configuration") +
  geom_label(data = ecs_stats, 
             aes(x = scenario, 
                 y = median + 0.4, 
                 label = paste(round(median, 2), "\n[", round(vlikely_lower, 2), ",", round(vlikely_upper, 2), "]")), 
             size = 3) +
  labs(y = "Weighted ECS Distributions (C)") +
  theme_light(base_size = 18) +
  theme(legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = "top",
        legend.position.inside = c(0.15, 0.89),
        legend.background = element_rect(fill = "white", colour = "lightgrey", linewidth = 0.5, linetype = "solid"),
        axis.title.x = element_blank(), 
        axis.text = element_text(size = 14), 
        axis.text.x = element_blank())

ecs_stats_plot
```
